refactor(moa): split gateway responsibilities
